cleanup publishing code
